lyra_clean/requirements.txt
# ============================================================================
# LYRA CLEAN - PRODUCTION DEPENDENCIES
# ============================================================================
# Cleaned from legacy: NO bert-score, rouge, nltk, scikit-learn, ispacenav
# Total size: ~150MB (was 650MB+)
# ============================================================================

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi>=0.104.0            # Modern async API framework
uvicorn[standard]>=0.24.0   # ASGI server with auto-reload
pydantic>=2.5.0             # Data validation + immutable models

# ============================================================================
# ASYNC DATABASE
# ============================================================================
aiosqlite>=0.19.0           # Async SQLite (replaces pandas+CSV)

# ============================================================================
# HTTP CLIENT (Async Ollama API)
# ============================================================================
httpx>=0.25.0               # Async HTTP (replaces requests)

# ============================================================================
# DATA PROCESSING (Minimal)
# ============================================================================
pandas>=2.0.0               # ONLY for CSV migration script (not runtime)

# ============================================================================
# CONFIGURATION
# ============================================================================
pyyaml>=6.0                 # Config files (optional, can use JSON)

# ============================================================================
# DEVELOPMENT (Optional)
# ============================================================================
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0      # Async test support
black>=23.0.0               # Code formatter
ruff>=0.1.0                 # Fast linter (replaces flake8+isort)

# ============================================================================
# NOTES
# ============================================================================
# REMOVED (from legacy):
# - bert-score (~200MB)      → No NLP metrics needed at runtime
# - rouge-score (~50MB)      → No NLP metrics needed at runtime
# - sacrebleu (~80MB)        → No NLP metrics needed at runtime
# - nltk (~100MB)            → Simple regex tokenization sufficient
# - scikit-learn (~80MB)     → No ML models needed
# - networkx (10MB)          → Graph in SQLite, not RAM
# - igraph (20MB)            → Graph in SQLite, not RAM
# - matplotlib (~60MB)       → No visualization at runtime
# - scipy (~50MB)            → No advanced math needed
#
# Total removed: ~650MB
# New total: ~150MB (77% reduction)
# ============================================================================
